move QuerySetUI into ui
